backend/user_api/pyproject.toml
[project]
name = "user-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "alembic>=1.17.1",
    "asyncpg>=0.30.0",
    "click>=8.3.0",
    "fastapi[standard]>=0.120.3",
    "faststream[kafka]>=0.6.3",
    "orjson>=3.11.4",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-jose[cryptography]>=3.5.0",
    "redis[hiredis]>=7.0.1",
    "sqlalchemy>=2.0.44",
    "tenacity>=9.1.2",
    "tldextract>=5.3.0",
]
